// src/app/core/model/user-especial.model.ts
export class UserEspecial {
  profileImg: string;
  othersImg1: string;
  othersImg2: string;
  othersImg3: string;
  first_name: string;
  last_name: string;
  email: string;
  checkImpairedFertility: boolean = false;
  checkBirthDisability: boolean = false;
  checkHaveTIIc: boolean = false;
  checkOvercomingExample: boolean = false;
  phone: number;
  phoneVisibility: string;
  profession: string;
  genre: string;
  orientation: string;
  aboutYou: string = '';
  cid: string[];
  surgeries: string[];
  hospitals: string[];
  medicines: string[];
}